+++ LeaguePackets/Game/165_S2C_TeamSurrenderStatus.cs
@@ -11,7 +11,7 @@ namespace LeaguePackets.Game
     public class S2C_TeamSurrenderStatus : GamePacket // 0xA5
     {
         public override GamePacketID ID => GamePacketID.S2C_TeamSurrenderStatus;
-        public uint SurrenderReason { get; set; }
+        public byte SurrenderReason { get; set; }
         public byte ForVote { get; set; }
         public byte AgainstVote { get; set; }
         public uint TeamID { get; set; }
@@ -19,14 +19,14 @@ namespace LeaguePackets.Game
         protected override void ReadBody(ByteReader reader)
         {
 
-            this.SurrenderReason = reader.ReadUInt32();
+            this.SurrenderReason = reader.ReadByte();
             this.ForVote = reader.ReadByte();
             this.AgainstVote = reader.ReadByte();
             this.TeamID = reader.ReadUInt32();
         }
         protected override void WriteBody(ByteWriter writer)
         {
-            writer.WriteUInt32(SurrenderReason);
+            writer.WriteByte(SurrenderReason);
             writer.WriteByte(ForVote);
             writer.WriteByte(AgainstVote);
             writer.WriteUInt32(TeamID);
